requirements.txt
marshmallow-sqlalchemy==0.25.0
fastapi==0.83.0
fastapi[uvicorn]
uvicorn==0.16.0
pydantic==1.8.2
databases==0.5.5
aiosqlite==0.17.0
